Added simple Session and User class
Session class to control the session and user-class to enable login and logout, created the user-table in the database too. Calls on the escape tring from the database and uses the methods in the database-class in the helper-folder to send in queries with find_by_query() and find_by_id().